+++ src/function.ts
@@ -108,17 +108,6 @@ export const getRing = <A, B>(R: Ring<B>): Ring<(a: A) => B> => {
   }
 }
 
-/**
- * Endomorphism form a monoid where the `empty` value is the identity function.
- *
- * @category instances
- * @since 2.10.0
- */
-export const getEndomorphismMonoid = <A = never>(): Monoid<Endomorphism<A>> => ({
-  concat: (x, y) => (a) => y(x(a)),
-  empty: identity
-})
-
 // -------------------------------------------------------------------------------------
 // utils
 // -------------------------------------------------------------------------------------
@@ -149,13 +138,6 @@ export interface Predicate<A> {
  */
 export interface Refinement<A, B extends A> {
   (a: A): a is B
-}
-
-/**
- * @since 2.0.0
- */
-export interface Endomorphism<A> {
-  (a: A): A
 }
 
 /**
@@ -758,3 +740,30 @@ export const hole: <T>() => T = absurd as any
  * @since 2.11.0
  */
 export const SK = <A, B>(_: A, b: B): B => b
+
+// -------------------------------------------------------------------------------------
+// deprecated
+// -------------------------------------------------------------------------------------
+
+/**
+ * Use `Endomorphism` module instead.
+ *
+ * @since 2.0.0
+ * @deprecated
+ */
+export interface Endomorphism<A> {
+  (a: A): A
+}
+
+/**
+ * Use `Endomorphism` module instead.
+ *
+ * @category instances
+ * @since 2.10.0
+ * @deprecated
+ */
+// tslint:disable-next-line: deprecation
+export const getEndomorphismMonoid = <A = never>(): Monoid<Endomorphism<A>> => ({
+  concat: (first, second) => flow(first, second),
+  empty: identity
+})